+++ examples/spring-data-graphql/src/main/java/com/blazebit/persistence/examples/spring/data/graphql/view/DogPetView.java
@@ -6,6 +6,7 @@
 package com.blazebit.persistence.examples.spring.data.graphql.view;
 
 import com.blazebit.persistence.examples.spring.data.graphql.model.Dog;
+import com.blazebit.persistence.integration.graphql.GraphQLDefaultFetch;
 import com.blazebit.persistence.view.EntityView;
 
 /**
@@ -15,6 +16,7 @@ import com.blazebit.persistence.view.EntityView;
 @EntityView(Dog.class)
 public interface DogPetView extends PetView {
 
+    @GraphQLDefaultFetch
     int getBarkCount();
 
 }